+++ NostalgiaOrbitDLL/Items/B01.cs
@@ -12,10 +12,10 @@ namespace NostalgiaOrbitDLL.Items
             ItemType = ItemTypes.B01;
 
             CanBuyUridium = true;
-            UridiumPurchase = new List<int>(2500);
+            UridiumPurchase = new List<float>(2500);
 
             CanSellByCredit = true;
-            CreditSale = new List<int>(10000);
+            CreditSale = new List<float>(10000);
 
             IsShield = true;
             Shield = 4000;
