+++ NostalgiaOrbitDLL/Items/G3N_3210.cs
@@ -12,10 +12,10 @@ namespace NostalgiaOrbitDLL.Items
             ItemType = ItemTypes.G3N_3210;
 
             CanBuyByCredit = true;
-            CreditPurchase = new List<int>(8000);
+            CreditPurchase = new List<float>(8000);
 
             CanSellByCredit = true;
-            CreditSale = new List<int>(2000);
+            CreditSale = new List<float>(2000);
 
             IsGear = true;
             GearSpeed = 4;
