+++ NostalgiaOrbitDLL/Items/G3N_6900.cs
@@ -12,10 +12,10 @@ namespace NostalgiaOrbitDLL.Items
             ItemType = ItemTypes.G3N_6900;
 
             CanBuyUridium = true;
-            UridiumPurchase = new List<int>(1000);
+            UridiumPurchase = new List<float>(1000);
 
             CanSellByCredit = true;
-            CreditSale = new List<int>(4000);
+            CreditSale = new List<float>(4000);
 
             IsGear = true;
             GearSpeed = 7;
