+++ NostalgiaOrbitDLL/ShopItem.cs
@@ -6,17 +6,17 @@ namespace NostalgiaOrbitDLL
     {
         // Credits
         public bool CanBuyByCredit { get; protected set; }
-        public List<int> CreditPurchase { get; protected set; }
+        public List<float> CreditPurchase { get; protected set; }
 
         public bool CanSellByCredit { get; protected set; }
-        public List<int> CreditSale { get; protected set; }
+        public List<float> CreditSale { get; protected set; }
 
         // Uridium
         public bool CanBuyUridium { get; protected set; }
-        public List<int> UridiumPurchase { get; protected set; }
+        public List<float> UridiumPurchase { get; protected set; }
 
         public bool CanSellUridium { get; protected set; }
-        public List<int> UridiumSale { get; protected set; }
+        public List<float> UridiumSale { get; protected set; }
 
         // Auction
         public bool CanBuyOnAuctionHouse { get; protected set; }
